refactor: extract registry admin cli commands
